Refactor of the repository for better clarity
