+++ global_config.py
@@ -3,10 +3,9 @@ from os.path import join, dirname, realpath
 
 SCRIPT_DIR = dirname(realpath(__file__))
 
-
-
-
 UPSCALE_SCRIPT_PATH = join(SCRIPT_DIR, "Real-ESRGAN", "inference_realesrgan.py")
-SEVEN_ZIP_PATH = join(SCRIPT_DIR, "7z2201-extra", "x64", "7za.exe")
+COMPRESS_METHOD = "7z" # "7z" or "integrated"
+SEVEN_ZIP_PATH = join(SCRIPT_DIR, "7z2201-extra", "x64", "7za.exe") # Path to 7z executable or None
 MODEL_NAME = "realesr-animevideov3"
+RAMDISK_LETTER = "R:" # The letter of the RAM disk (Windows only, None for no RAM disk)
 VERSION = "0.0.1"
