+++ upscale.py
@@ -1,51 +1,27 @@
 import argparse
 import glob
 import os
-from pathlib import Path
 import re
 import subprocess
 import sys
-import zipfile
 from PIL import Image
 
 Image.MAX_IMAGE_PIXELS = None
 
 import multiprocessing
-import threading # For IO
-import shutil
 import time
 
-import rich
-import rich.progress
 from rich.progress import track, Progress
 # from rich import print
 from rich.console import Console
 
 import natsort
-from paths import OutputPathGenerator, PathGenerator
+from paths import OutputPathGenerator
 from global_config import *
+from utils.files import get_input_dir, rm_tree, prune_empty_folders
+from utils.compression import compress, extract
 
 console = Console()
-
-# if __name__ == '__main__':
-#     # Call subprocess to ask for the version of python, opencv and torch (avoid loading them in the main process to avoid high memory usage)
-#     p = [
-#         subprocess.Popen([sys.executable, "--version"], stdout=subprocess.PIPE),
-#         subprocess.Popen([sys.executable, "-c", "import cv2; print(f'cv2={cv2.__version__}')"], stdout=subprocess.PIPE),
-#         subprocess.Popen([sys.executable, "-c", "import torch; print(f'torch={torch.__version__}')"], stdout=subprocess.PIPE)
-#     ]
-#     for process in p:
-#         process.wait()
-#         print(process.stdout.read().decode('utf-8').strip('\n'))
-
-
-
-# pyjion
-# if __name__ == '__main__': print(f"Importing pyjion...")
-# import pyjion
-# pyjion.config(level=1)
-# if __name__ == '__main__': print(f"Pyjion: {pyjion.enable()}")
-
 
 def parse_args():
     parser = argparse.ArgumentParser()
@@ -103,39 +79,6 @@ def get_cbz_files(input_path, ignore_upscaled=False):
 
     return res
 
-def extract(file, output_path=None, progress=Progress(transient=True)):
-    # Create a folder with the same name as the file
-    folder_name = file[:-4] if output_path is None else output_path
-    folder_name = os.path.abspath(folder_name)
-    folder_name = os.path.normpath(folder_name)
-    if os.path.exists(folder_name):
-        # Empty the folder
-        rm_tree(folder_name)
-
-    if not os.path.exists(folder_name):
-        os.mkdir(folder_name)
-
-    # Extract the files to the folder
-    # task_id = progress.add_task(f"Extracting {file}...", total=os.path.getsize(file))
-    # add tabulation before the message to align it with the other messages
-    task_id = progress.add_task(f"    Extracting {file}...", total=os.path.getsize(file))
-    p = progress.open(file, "rb", task_id=task_id)
-    with zipfile.ZipFile(p) as zip_ref:
-        zip_ref.extractall(folder_name)
-    p.close()
-    progress.remove_task(task_id)
-    # Move the files to the root of the folder
-    for file in glob.glob(folder_name + '/*/**', recursive=True):
-        if os.path.isfile(file):
-            os.rename(file, folder_name + '/' + os.path.basename(file))
-
-    # Remove the empty folders
-    for folder in glob.glob(folder_name + '/*/**', recursive=True):
-        if os.path.isdir(folder):
-            os.rmdir(folder)
-
-    return folder_name
-
 def divide(input_folder, output_folder, divide, format=None):
     if divide <= 1:
         return
@@ -188,61 +131,6 @@ def fit_to_width(input_folder, width, format=None):
         print(f"  {i + 1}/{size} {os.path.basename(image)}", end='\r')
 
 
-def compress_seven_zip(folder, output_path, progress=Progress(transient=True)):
-    # Compress the folder to a cbz file
-    if not os.path.exists(SEVEN_ZIP_PATH):
-        raise Exception(f"7z.exe not found at {SEVEN_ZIP_PATH}")
-
-    task_id = progress.add_task(f"    Compressing {folder}...", total=None)
-    ret = subprocess.run([SEVEN_ZIP_PATH, "-tzip", 'a', output_path, folder + '/*'], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-    progress.remove_task(task_id)
-    if ret.returncode != 0:
-        raise Exception(f"Failed to compress {folder} to {output_path}")
-
-def get_size(folder):
-    size = 0
-    for file in glob.glob(folder + '/**', recursive=True):
-        if os.path.isdir(file):
-            continue
-        size += os.path.getsize(file)
-    return size
-
-def compress_integrated(folder, output_path, progress=Progress(transient=True)):
-    """
-    Compress the folder to a cbz file using standard python libraries
-
-    Args:
-        folder (str): The folder to compress
-        output_path (str): The path to the output file
-        progress (Progress): The progress bar to use
-    """
-
-    # Compress the folder to a cbz file using standard python libraries
-    # Integraded version seems to have trouble with network drives, use 7zip in those cases
-    task_id = progress.add_task(f"    Compressing {folder}...", total=get_size(folder))
-
-    with zipfile.ZipFile(output_path, 'w', zipfile.ZIP_DEFLATED) as zip_ref:
-        for file in glob.glob(folder + '/**', recursive=True):
-            if os.path.isdir(file):
-                continue
-            zip_ref.write(file, os.path.relpath(file, folder))
-            progress.update(task_id, advance=os.path.getsize(file))
-
-    progress.remove_task(task_id)
-
-def compress(folder, output_path):
-    """
-    Compress the folder to a cbz file and select the best method
-
-    Args:
-        folder (str): The folder to compress
-        output_path (str): The path to the output file
-    """
-    os.makedirs(os.path.dirname(output_path), exist_ok=True)
-    if os.path.exists(SEVEN_ZIP_PATH):
-        compress_seven_zip(folder, output_path)
-    else:
-        compress_integrated(folder, output_path)
 
 def upscale(input_path, output_path, scale, format=None, width=0, tiles=1024, wait=True, tile_pad=50, fp32=False):
     args = ["python", UPSCALE_SCRIPT_PATH, '-i', input_path, '-o', output_path, '-n', MODEL_NAME, '-s', f"{scale}", "-t", f"{tiles}", "--tile_pad", f"{tile_pad}", "--width", f"{width}"]
@@ -258,14 +146,6 @@ def upscale(input_path, output_path, scale, format=None, width=0, tiles=1024, wa
     if wait:
         ret.wait()
 
-    # end_after_upscaling = (False)
-    # while ret.poll() is None:
-    #     # Check if the user has pressed q to quit after the upscaling
-    #     if keyboard.is_pressed('shift') and keyboard.is_pressed('q') and keyboard.is_pressed('ctrl') and not end_after_upscaling:
-    #         print("<<< User pressed ctrl + q, no more files will be processed after the current one >>>")
-    #         end_after_upscaling = True
-    #     time.sleep(0.1)
-
     if wait and ret.returncode != 0:
         raise Exception("Upscaling failed")
 
@@ -279,41 +159,6 @@ def upscale(input_path, output_path, scale, format=None, width=0, tiles=1024, wa
         print(f"WARNING: Input length ({input_length}) inferior to output length ({output_length}). Could be due to an image being split into multiple images.")
 
     return ret
-
-def upscale_parallel(input_path, output_path, scale, format=None, width=0, tiles=1024, thread_count=2):
-    # Get all the files in the input path
-    files = glob.glob(input_path + '/**', recursive=True)
-    files = [f for f in files if not os.path.isdir(f)]
-    if format is not None:
-        files = [f for f in files if f.endswith(f'.{format}')]
-
-    # Split the files into chunks
-    subprocesses = []
-    while len(files) > 0:
-        if len(subprocesses) >= thread_count:
-            # Wait for a thread to finish
-            for t in subprocesses:
-                if t.poll() is not None:
-                    subprocesses.remove(t)
-                    break
-            time.sleep(0.1)
-        else:
-            f = files[0]
-            print(f"Upscaling {f} to {f.replace(input_path, output_path)}")
-            # threads.append(threading.Thread(target=upscale, args=(f, output_path, scale, format, width, tiles)))
-            subprocesses.append(upscale(f, output_path, scale, format, width, tiles, wait=False, tile_pad=args.tile_pad, fp32=args.fp32))
-            # upscale(f, output_path, scale, format, width, tiles, wait=True)
-            files.pop(0)
-
-
-def rm_tree(path):
-    try:
-        shutil.rmtree(path)
-    except OSError as e:
-        print("Error: %s : %s" % (path, e.strerror))
-
-def normalize_path(path):
-    return os.path.abspath(os.path.normpath(path))
 
 def sync_files(args, file_mapping: dict):
     # Remove files that have been upscaled but not in the input folder anymore
@@ -338,34 +183,6 @@ def sync_files(args, file_mapping: dict):
                     os.remove(file)
                 except:
                     print(f"    <<< Error removing {file} >>>")
-
-
-def prune_empty_folders(path):
-    # threads = []
-    for root, dirs, files in track(os.walk(path, topdown=False), description="Removing empty folders...", transient=True):
-        for name in dirs:
-            folder = os.path.join(root, name)
-            try:
-                if len(os.listdir(folder)) == 0:
-                    print(f"Removing empty folder {folder}")
-                    os.rmdir(folder)
-                    # threads.append(threading.Thread(target=os.rmdir, args=(folder,)))
-            except OSError as e:
-                print("Error: %s : %s" % (folder, e.strerror))
-
-    # for t in threads:
-    #     t.start()
-
-    # # Wait for all threads to finish
-    # for t in track(threads, description="Removing empty folders...", transient=True):
-    #     t.join()
-
-
-def get_input_dir(input_path):
-    input_directory = os.path.abspath(input_path)
-    if not os.path.isdir(input_path):
-        input_directory = os.path.dirname(input_directory)
-    return input_directory
 
 
 def main(args, params, file_mapping: dict):
@@ -452,12 +269,6 @@ def main(args, params, file_mapping: dict):
 
     return len([path for path in paths if path is not None])
 
-# def print_sleep(duration: int, step: int = 5):
-#     for i in range(0, duration, step):
-#         print(f"\033[KSleeping for {duration - i} seconds... ", end='\r')
-#         time.sleep(step)
-#     print("\033[K", end='\r')
-
 def print_sleep(duration: int, step: int = 0.5):
     # Use rich to print sleep
     with Progress(speed_estimate_period=2, transient=True) as progress:
@@ -468,7 +279,6 @@ def print_sleep(duration: int, step: int = 0.5):
             time.sleep(step)
         time.sleep(duration % step)
         progress.stop()
-
 
 def start_processing(args, params):
     file_mapping: dict = params.file_mapping
