+++ utils/paths.py
@@ -1,0 +1,25 @@
+import os
+import subprocess
+
+def get_program_path(program_name):
+    """
+    Get the path to the program with the given name
+
+    Args:
+        program_name (str): The name of the program
+
+    Returns:
+        str: The path to the program
+    """
+    try:
+        # Windows
+        if os.name == 'nt':
+            return subprocess.check_output(['where', program_name]).decode().strip()
+        # Linux and MacOS
+        else:
+            return subprocess.check_output(['which', program_name]).decode().strip()
+    except subprocess.CalledProcessError:
+        return None
+
+def normalize_path(path):
+    return os.path.abspath(os.path.normpath(path))
